chore: make package work as either required, or called directly
