+++ package.json
@@ -3,6 +3,7 @@
   "version": "1.0.8",
   "description": "Convert EML files to PDF, save attachments, rename EML files",
   "main": "index.js",
+  "bin": "index.js",
   "dependencies": {
     "dateformat": "^3.0.3",
     "envelope": "^2.0.2",
